# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: openGym API
  version: 1.3.9
  summary: Passkey (WebAuthn) auth + per-user state storage for openGym
  description: |
    The backend of [openGym](https://opengym.duarte-santos.ch) — a free, open-source
    gym & body-weight tracker. One Node file, no framework, JSON-file storage,
    HMAC-signed session cookies.

    ### Authentication
    Sign-in is by **passkey** (WebAuthn, via `@simplewebauthn/server`). A successful
    `register/verify` or `login/verify` sets the session cookie; every authenticated
    route then works with that cookie. The paired mobile app has no shared cookie jar,
    so it redeems a short pairing code (`/api/pair/redeem`) for the *same* signed token
    and sends it as `Authorization: Bearer <token>` instead.

    An instance started with `PASSWORD_LOGIN=1` also offers **name-and-password
    sign-in** (tag `password`): opt-in per profile, scrypt-hashed, throttled, with an
    admin-issued one-time reset code instead of e-mail. A profile may also add an e-mail
    address to sign in with instead of its name (`/api/account/email`); no mail is ever
    sent to it. Its routes set exactly the same cookie. With the flag off (the default)
    every one of them answers 404.

    ### Throttling
    The `password` routes are counted per client address: at most 60 requests a minute
    across them, and wrong answers (a password, a reset code, an invite code on password
...

    ### Sessions
...
    the account.

    ### CSRF
...

...

...
  license:
...
servers:
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /api/challenges:
    get:
      tags: [challenges]
      operationId: getChallenges
      summary: List the signed-in profile's challenges
      responses:
        '200':
          description: Challenge summaries with scores calculated from each participant's private state.
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [challenges]
      operationId: postChallenge
      summary: Create a two-person challenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, startDate, endDate, workoutsPerWeek, startWeight, targetWeight]
              properties:
                name: { type: string, maxLength: 60 }
                startDate: { type: string, format: date }
                endDate: { type: string, format: date }
                workoutsPerWeek: { type: integer, minimum: 1, maximum: 14 }
                stake: { type: string, maxLength: 160 }
                startWeight: { type: number, minimum: 25, maximum: 400 }
                targetWeight: { type: number, minimum: 25, maximum: 400 }
                unit: { type: string }
                timeZone: { type: string }
      responses:
        '201': { description: Created, with a one-time invitation code in the challenge. }
        '400': { description: Invalid challenge fields. }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/challenges/detail:
    get:
      tags: [challenges]
      operationId: getChallengeDetail
      summary: Read one challenge and its current versus score
      parameters:
        - in: query
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: Challenge detail. Exact weights are included only for the caller. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { description: Missing challenge or caller is not a participant. }
  /api/challenges/join:
    post:
      tags: [challenges]
      operationId: postChallengeJoin
      summary: Join a challenge using its invitation code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, startWeight, targetWeight]
              properties:
                code: { type: string }
                startWeight: { type: number, minimum: 25, maximum: 400 }
                targetWeight: { type: number, minimum: 25, maximum: 400 }
                unit: { type: string }
      responses:
        '200': { description: Joined. The invitation code is now closed. }
        '400': { description: Invalid weights. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { description: Invitation code not found. }
        '409': { description: Caller already joined or the challenge is full. }
  /api/challenges/photo:
    get:
      tags: [challenges]
      operationId: getChallengePhoto
      summary: Read a participant photo
      description: Only participants in that challenge may read the file.
      parameters:
        - in: query
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: user
          required: true
          schema: { type: string }
      responses:
        '200':
          description: JPEG, PNG or WebP image bytes.
          content:
            image/jpeg: { schema: { type: string, format: binary } }
            image/png: { schema: { type: string, format: binary } }
            image/webp: { schema: { type: string, format: binary } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { description: Missing photo, challenge, or membership. }
    post:
      tags: [challenges]
      operationId: postChallengePhoto
      summary: Set the caller's square versus photo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, photo]
              properties:
                id: { type: string }
                photo: { type: string, description: JPEG, PNG or WebP data URL; decoded maximum 750 KiB. }
      responses:
        '200': { description: Stored on the persistent data volume. }
        '400': { description: Invalid image data. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { description: Missing challenge or caller is not a participant. }
        '413': { description: Photo is over 750 KiB after decoding. }
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: gymsid
      description: |
        HMAC-signed session cookie, set by `register/verify` / `login/verify`.
        On HTTPS deployments it is issued as **`__Host-gymsid`** (host-locked,
        Secure); the unprefixed `gymsid` is what a plain-http `localhost` instance
        uses, and is still accepted for pre-upgrade sessions. HttpOnly + SameSite=Lax;
        browser clients never handle the value directly.
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        The same signed session token, carried as `Authorization: Bearer <token>` by
        the paired mobile app (obtained from `POST /api/pair/redeem`). Bearer requests
        skip the CSRF origin check — a browser never attaches this header on its own.
  responses:
    Unauthorized:
      description: No session, expired/revoked session, or a disabled account.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: "not signed in" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable message. This is the only error shape the API has.
        code:
          type: string
          description: >-
            A stable reason on the password, throttle and media routes, for a client to word
            in its own language (`bad-credentials`, `locked`, `too-short`, `media-quota`, …).
            Media refusals add the numbers to word it with (`maxMB`, `maxSec`, `usedMB`,
            `quotaMB`).
        retryAfter:
          type: integer
          description: 429 only — the same number of seconds as the Retry-After header.
...